refactor: 공통 로직 utils로 이동
