+++ src/pages/OrderPage.tsx
@@ -3,6 +3,7 @@ import { OrderList } from '../component/OrderList';
 import Pagenation from '../component/Pagenation';
 import { useOrderData } from '../hooks/useOrderData';
 import { usePagenation } from '../hooks/usePagenation';
+import { splitArray } from '../utils/splitArray';
 
 export default function OrderPage() {
   const api = new MockApi();
@@ -10,7 +11,7 @@ export default function OrderPage() {
 
   const [data] = useOrderData(api, checkDate);
   const [pageNumber] = usePagenation({ total: data?.length || 0, size: 50 });
-  const pageData = splitIntoChunk(data, 50);
+  const pageData = splitArray(data, 50);
 
   return (
     <div>
@@ -18,19 +19,4 @@ export default function OrderPage() {
       <Pagenation total={data?.length || 0} size={50} />
     </div>
   );
-}
-
-function splitIntoChunk(arr: any, chunk: number) {
-  if (arr === undefined) {
-    return [];
-  }
-
-  const result = [];
-
-  for (let index = 0; index < arr.length; index += chunk) {
-    const tempArray = arr.slice(index, index + chunk);
-    result.push(tempArray);
-  }
-
-  return result;
 }
